Work in a temporary directory
`cookiecutter` is now instructed to create the new project as a
subdirectory of a new temporary directory. A `src` folder is created and
the specified contents are moved there; then everything is cleaned up.

A temporary directory is necessary in case the chosen project name
conflicts with a folder supplied for the initial `src`.