Use proper licenses
The appropriate license text has been copied in from choosealicense.com
and adapted to use the correct placeholders (for the MIT Licence, which
includes an explicit statement of copyright). The files have also been
renamed in anticipation of the next change, so the project is
temporarily broken, but it still seemed better to split the commit.