Consistent license naming
Licenses are now referred to by their name as they will appear in the
prose statement at the end of the README (without the word "the"). This
should be more maintainable in the long run.

Unfortunately there doesn't seem to be a way to make `cookiecutter` use
the names in the `templates` directory as a list of options for the
license selection in the UI - it will have to be updated separately.