Add gitignores and local scratch file setup
The cookiecutter now adds suitable `.gitignore` files and a `.local`
folder to new projects. The `.local` folder is set up with a do-nothing
`run-acceptance-test` script and a nested `.gitignore` that hides the
folder's contents from Git, without the main `.gitignore` needing to
mention the folder's existence.

So that the `.gitignore` files in the cookiecutter are not processed by
Git for the overall `cookiebaker` repository, they're given special
names for template substitution (just like the main folder for the
template). The idea is that their name isn't literally `.gitignore`, but
they'll correspond to files that do have that name in the created
projects.